happy

+++ onvm/onvm_nflib/onvm_framework.c
@@ -194,8 +194,8 @@ onvm_framework_cpu(int thread_id)
 
 		// rx
 		do {
-			if (BATCH_SIZE != cl->batch_size) {
-				BATCH_SIZE = cl->batch_size;
+			if (BATCH_SIZE != (int)cl->batch_size) {
+				BATCH_SIZE = (int)cl->batch_size;
 				RTE_LOG(INFO, APP, "Batch size changed to %d\n", BATCH_SIZE);
 			}
 			num_packets = rte_ring_dequeue_bulk(rx_q, (void **)batch->pkt_ptr[buf_id], BATCH_SIZE, NULL);
